+++ plugins/e57/CMakeLists.txt
@@ -5,10 +5,10 @@ add_subdirectory(libE57Format)
 
 PDAL_ADD_PLUGIN(e57plugin_reader reader e57
         FILES
-            io/e57reader.cpp
-            io/scan.cpp
-            io/utils.cpp
-            io/uuid.cpp
+            io/E57Reader.cpp
+            io/Scan.cpp
+            io/Utils.cpp
+            io/Uuid.cpp
         LINK_WITH
             E57Format
             ${XercesC_LIBRARY}
@@ -18,10 +18,10 @@ target_include_directories(${e57plugin_reader} PRIVATE "libE57Format/include")
 
 PDAL_ADD_PLUGIN(e57plugin_writer writer e57
         FILES
-            io/e57writer.cpp
-            io/scan.cpp
-            io/utils.cpp
-            io/uuid.cpp
+            io/E57Writer.cpp
+            io/Scan.cpp
+            io/Utils.cpp
+            io/Uuid.cpp
         LINK_WITH
             E57Format
             ${XercesC_LIBRARY}
@@ -32,9 +32,9 @@ target_include_directories(${e57plugin_writer} PRIVATE "libE57Format/include")
 if (WITH_TESTS)
     PDAL_ADD_TEST(pdal_io_e57_reader_test
             FILES
-                test/reader-test.cpp
-                test/scan-test.cpp
-                test/utils-test.cpp
+                test/E57ReaderTest.cpp
+                test/ScanTest.cpp
+                test/E57UtilsTest.cpp
             LINK_WITH
                 ${e57plugin_reader}
                 E57Format
@@ -43,8 +43,8 @@ if (WITH_TESTS)
 
     PDAL_ADD_TEST(pdal_io_e57_writer_test
             FILES
-                test/writer-test.cpp
-                test/uuid-test.cpp
+                test/E57WriterTest.cpp
+                test/E57UuidTest.cpp
             LINK_WITH
                 ${e57plugin_reader}
                 ${e57plugin_writer}
