+++ plugins/e57/io/E57Writer.hpp
