+++ plugins/e57/io/Utils.hpp
@@ -32,48 +32,38 @@
 * OF SUCH DAMAGE.
 ****************************************************************************/
 
-#include <gtest/gtest.h>
-#include "Support.hpp"
+#pragma once
 
-#include "../io/scan.hpp"
-#include "../io/e57reader.hpp"
+#include <string>
+#include <E57Format.h>
+#include <pdal/Dimension.hpp>
 
-using namespace pdal;
-
-TEST(ScanTest,testScanCtr) 
+namespace pdal
 {
-    E57Reader reader(Support::datapath("e57/A4.e57"));
-    auto scans = reader.getScans();
-    auto firstScan = scans[0];
-    ASSERT_EQ(firstScan->getNumPoints(),(pdal::point_count_t)4);
+namespace e57plugin
+{
+    // converts an e57 dimension string to a pdal dimension
+    // returns pdal::Dimension::Id::Unknown in case the dimension is not recognised
+    pdal::Dimension::Id e57ToPdal(std::string e57Dimension);
+
+    // converts a pdal dimension to the corresponding E57 string
+    // returns an empty string in case the dimension is not recognised
+    std::string pdalToE57(pdal::Dimension::Id pdalDimension);
+
+    /// Converts a value from E57 to pdal. Handles change in type representation
+    /// For example, intensity in e57 is between 0 and 1 and 0 and 2^16 in pdal
+    double rescaleE57ToPdalValue(std::string e57Dimension, double value,const std::pair<double,double>& e57Bounds);
+
+    std::vector<pdal::Dimension::Id> supportedPdalTypes();
+    std::vector<std::string> supportedE57Types();
+
+    // Tries to find the limit of a dimension in the e57 node headers
+    // return nan if not found
+    std::pair<double,double> getLimits(const e57::StructureNode &prototype, std::string fieldName);
+
+    std::string getDescription(const e57::Node &node);
+
+    // Get the bounds of a given dimension as expected by pdal
+    std::pair<double,double> getPdalBounds(pdal::Dimension::Id id);
 }
-
-TEST(ScanTest,getDimension) 
-{
-    E57Reader reader(Support::datapath("e57/A4.e57"));
-    auto scans = reader.getScans();
-    auto firstscan = scans[0];
-
-    auto dimensions = firstscan->getDimensions();
-    ASSERT_EQ(dimensions.size(),(unsigned long)7);
-    std::vector<std::string> expectedDimensions = {"cartesianX","cartesianY","cartesianZ",
-        "colorRed","colorGreen","colorBlue","intensity"};
-    for (auto dim: expectedDimensions)
-    {
-        ASSERT_TRUE(dimensions.find(dim) != dimensions.end());
-    }
-}
-
-TEST(ScanTest,testGetPoints) 
-{
-    E57Reader reader(Support::datapath("e57/A_B.e57"));
-    auto scans = reader.getScans();
-    auto firstScan = scans[0];
-    auto pts = firstScan->getPoints();
-    ASSERT_EQ((pdal::point_count_t)pts.childCount(),firstScan->getNumPoints());
-    ASSERT_EQ(pts.childCount(),2);
-    auto secondScan = scans[1];
-    pts = secondScan->getPoints();
-    ASSERT_EQ((pdal::point_count_t)pts.childCount(),secondScan->getNumPoints());
-    ASSERT_EQ(pts.childCount(),(int64_t)4);
 }
